chore: cleanup and refine
